Split tests into different modules - revise setup
- you can now use the `makeTestFactory` setup function to create an `it` function which will auto-setup and pass down the WebDriver, and also deal with ensuring it quits after the test is done
- also renamed these tests from `.integration-test.` to `.browser-test.` to be more descriptive
